requirements.txt
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
Requests==2.30.0
selenium==4.9.1
webdriver_manager==3.8.5
